refactor: ページのファイル配置を URL 構造に合わせる
ui/Dashboard/ や app/dashboard/ という名前は担当している URL と対応して
おらず、どのファイルがどのページかを名前から辿れなかった。

  ui/Dashboard/index.tsx    → ui/pages/home.tsx              "/"
  ui/DeviceDetail/index.tsx → ui/pages/devices/details.tsx   "/devices/:device"
  app/dashboard/index.tsx   → app/home.tsx                   "/"
  app/devices/index.tsx     → app/devices/details.tsx        "/devices/:device"

app/ と ui/pages/ がページ単位で 1:1 に対応する。マウントパスは変えて
いないので AppType の形は不変で、hc クライアントに影響しない。

移動と import パスの調整のみで、ロジックの変更はない。
